🚚 [variables] Move module `variables` to `domain`
