+++ src/cutty/variables/domain/variables.py
