+++ tests/unit/templates/domain/test_render.py
@@ -7,7 +7,7 @@ from cutty.filestorage.domain.files import SymbolicLink
 from cutty.filesystems.domain.purepath import PurePath
 from cutty.templates.domain.render import Renderer
 from cutty.variables.bindings import Binding
-from cutty.variables.variables import GenericVariable
+from cutty.variables.domain.variables import GenericVariable
 
 
 @pytest.mark.parametrize(
