+++ tests/unit/variables/test_prompts.py
@@ -10,8 +10,8 @@ from prompt_toolkit.output import DummyOutput
 
 from cutty.variables.binders import Binder
 from cutty.variables.bindings import GenericBinding
+from cutty.variables.domain.variables import GenericVariable
 from cutty.variables.prompts import createprompt
-from cutty.variables.variables import GenericVariable
 from tests.util.keys import BACKSPACE
 from tests.util.keys import ESCAPE
 from tests.util.keys import RETURN
